requirements.txt

flask
pytz
requests
pandas
python-telegram-bot==13.15
